third_party/WebKit/LayoutTests/editing/inserting/insert_and_place_holder.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>|<br></div>',
    'insertHTML abc<div>def</div>',
    '<div contenteditable>abc<div>def|</div></div>'),
    'displaced generic placeholder');

// Note: Blink doesn't use class="webkit-block-placeholder" anymore.
test(() => assert_selection(
    '<div contenteditable>|<br class="webkit-block-placeholder"></div>',
    'insertHTML abc<div>def</div>',
    '<div contenteditable>abc<div>def|</div></div>'),
    'displaced placeholder with "class" attribute');
</script>
